requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Bcrypt
Flask-Login
PyJWT
python-dotenv